+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,23 @@
 --->
 
 ## Upcoming Release
+* Fixed parameters swap in VWan HubVnet connection
+* Added new cmdlets for Azure Network Virtual Appliance Sites
+    - `Get-AzVirtualApplianceSite`
+    - `New-AzVirtualApplianceSite`
+    - `Remove-AzVirtualApplianceSite`
+    - `Update-AzVirtualApplianceSite`
+    - `New-AzOffice365PolicyProperty`
+* Added new cmdlets for Azure Network Virtual Appliance
+    - `Get-AzNetworkVirtualAppliance`
+    - `New-AzNetworkVirtualAppliance`
+    - `Remove-AzNetworkVirtualAppliance`
+    - `Update-AzNetworkVirtualAppliance`
+    - `Get-AzNetworkVirtualApplianceSku`
+    - `New-AzVirtualApplianceSkuProperty`
+* Onboarded Application Gateway to Private Link Common Cmdlets
+* Onboarded StorageSync to Private Link Common Cmdlets
+* Onboarded SignalR to Private Link Common Cmdlets
 
 ## Version 3.1.0
 * Added support for AddressPrefixType parameter to `Remove-AzExpressRouteCircuitConnectionConfig`